% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmld.R
\name{downloadPopByChr}
\alias{downloadPopByChr}
\title{download hapmap resource with LD estimates}
\usage{
downloadPopByChr(
  chrname = "chr1",
  popname = "CEU",
 
    urlTemplate = "http://hapmap.ncbi.nlm.nih.gov/downloads/ld_data/2009-02_phaseIII_r2/ld_\%\%CHRN\%\%_\%\%POPN\%\%.txt.gz",
  targfolder = Sys.getenv("LDBLOCK_TXTGZ_DIR")
)
}
\arguments{
\item{chrname}{UCSC format tag for chromosome}

\item{popname}{hapmap three letter code for population, e.g. 'CEU'}

\item{urlTemplate}{pattern for creating URL given chr and pop}

\item{targfolder}{destination}
}
\value{
just run for side effect of download.file
}
\description{
download hapmap resource with LD estimates
}
\details{
delivers HapMap LD data to `targfolder`
}
\examples{

\dontrun{
 downloadPopByChr()
 }

}
\keyword{models}
