% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUR_singletons}
\alias{EUR_singletons}
\title{singletons from EUR}
\format{
character vector
}
\source{
\url{ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/working/20130606_sample_info/20130606_sample_info.xlsx}, to which superpopulation codes were added
}
\usage{
EUR_singletons
}
\description{
singletons from EUR
}
\keyword{datasets}
