\name{ldblock-package}
\alias{ldblock-package}
\alias{ldblock}
\docType{package}
\title{
\packageTitle{ldblock}
}
\description{
\packageDescription{ldblock}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ldblock}
\packageIndices{ldblock}


}
\author{
\packageAuthor{ldblock}

Maintainer: \packageMaintainer{ldblock}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
# see vignette
}
