% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmld.R
\name{ldmat,ldstruct-method}
\alias{ldmat,ldstruct-method}
\title{accessor for matrix component}
\usage{
\S4method{ldmat}{ldstruct}(x)
}
\arguments{
\item{x}{instance of ldstruct}
}
\description{
accessor for matrix component
}
