% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldlink.R
\name{ldmat}
\alias{ldmat}
\title{use LDmat API from NCI LDlink service}
\usage{
ldmat(rsvec, pop = "CEU", type = "d",
  token = Sys.getenv("LDLINK_TOKEN"))
}
\arguments{
\item{rsvec}{character vector of SNP ids}

\item{pop}{three letter code for HapMap population, defaults to CEU}

\item{type}{'r2' or 'd', defaults to 'd' implying d-prime}

\item{token}{the API token provided by NCI, defaults to value of environment variable LDLINK_TOKEN}
}
\value{
data.frame
}
\description{
use LDmat API from NCI LDlink service
}
\examples{
if (interactive()) ldmat(c("rs77749396","rs9303279","rs9303280","rs9303281"))
}
