% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampinf_1kg}
\alias{sampinf_1kg}
\title{population and relationship information for 1000 genomes}
\format{
data.frame
}
\source{
\url{ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/working/20130606_sample_info/20130606_sample_info.xlsx}, to which superpopulation codes were added
}
\usage{
sampinf_1kg
}
\description{
population and relationship information for 1000 genomes
}
\keyword{datasets}
