% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_1kg.R
\name{stack1kg}
\alias{stack1kg}
\title{couple together a group of VCFs}
\usage{
stack1kg(chrs = as.character(1:22), index = FALSE, useEBI = TRUE)
}
\arguments{
\item{chrs}{a vector of chromosome names for extraction from 1000 genomes
VCF collection}

\item{index}{logical telling whether VcfStack should attempt to create the
local index; for 1000 genomes, the tbi are in the cloud and will be used by
readVcf so FALSE is appropriate}

\item{useEBI}{logical(1) defaults to TRUE ... use tabix-indexed vcf from EBI}
}
\description{
couple together a group of VCFs
}
