% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldblock-package.R
\docType{package}
\name{ldblock-package}
\alias{ldblock-package}
\alias{ldblock}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"ldblock")\Sexpr{tools:::Rd_package_title("ldblock")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"ldblock")\Sexpr{tools:::Rd_package_description("ldblock")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"ldblock")\Sexpr{tools:::Rd_package_DESCRIPTION("ldblock")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"ldblock")\Sexpr{tools:::Rd_package_indices("ldblock")}
}
\examples{

# see vignette

}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"ldblock")\Sexpr{tools:::Rd_package_author("ldblock")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"ldblock")\Sexpr{tools:::Rd_package_maintainer("ldblock")}
}
\keyword{package}
