% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldblock-package.R
\docType{class}
\name{ldstruct-class}
\alias{ldstruct-class}
\title{Class \code{"ldstruct"}}
\description{
Manage information about LD statistics as reported by HapMap.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ldstruct", ...)}.
}

\examples{

showClass("ldstruct")

}
\keyword{classes}
